+++ dto/src/main/java/ca/on/oicr/gsi/runscanner/dto/PacBioNotificationDto.java
@@ -1,6 +1,9 @@
 package ca.on.oicr.gsi.runscanner.dto;
 
 import ca.on.oicr.gsi.runscanner.dto.type.Platform;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
@@ -8,6 +11,12 @@ import java.util.Optional;
 public class PacBioNotificationDto extends NotificationDto {
 
   private Map<String, String> poolNames;
+
+  public static record SMRTCellPosition(
+      String position, String containerSerialNumber, String poolName, String movieLength) {}
+
+  @JsonInclude(Include.NON_NULL)
+  private List<SMRTCellPosition> positionList;
 
   @Override
   public boolean equals(Object obj) {
@@ -36,6 +45,10 @@ public class PacBioNotificationDto extends NotificationDto {
     return poolNames;
   }
 
+  public List<SMRTCellPosition> getPositionList() {
+    return positionList;
+  }
+
   @Override
   public int hashCode() {
     return Objects.hash(super.hashCode(), poolNames);
@@ -45,8 +58,17 @@ public class PacBioNotificationDto extends NotificationDto {
     this.poolNames = poolNames;
   }
 
+  public void setPositionList(List<SMRTCellPosition> positionList) {
+    this.positionList = positionList;
+  }
+
   @Override
   public String toString() {
-    return super.toString() + ", PacBioNotificationDto [poolNames = " + poolNames + "]";
+    return super.toString()
+        + ", PacBioNotificationDto [poolNames= "
+        + poolNames
+        + ", positionList= "
+        + positionList
+        + "]";
   }
 }